Some errors handled

+++ vkbot.py
@@ -69,6 +69,7 @@ class VkBot:
     def user_command_handler(self, event: vk_api.bot_longpoll.VkBotMessageEvent, db: DataBase) -> None:
         # TODO: Add errors handler
         # TODO: DELETE DEBUG INFO
+        # TODO: Warning: unexpected arguments appears when count of functions more than 10 ????
         command = (event.obj["message"]["text"].split())[0]
         commands_dict = {
             '/add_author': self.add_author,
@@ -101,7 +102,6 @@ class VkBot:
         _, author_personal_id = (event.obj["message"]["text"].split())
         author_id = self.vk_session.method(method="users.get", values={"user_ids": author_personal_id})[0]["id"]
 
-        # check if user is admin
         if db.user_role_check(event.obj["message"]["from_id"], "admin") and \
                 not db.user_role_check(author_id, "author"):
             db.sql_execute_query(f'INSERT INTO author VALUES ({author_id})')
@@ -116,7 +116,6 @@ class VkBot:
                                   user_id=event.obj["message"]["from_id"])
 
     def get_authors(self, event: vk_api.bot_longpoll.VkBotMessageEvent, db: DataBase) -> None:
-        # check if user is admin
         if db.user_role_check(event.obj["message"]["from_id"], "admin"):
             self.send_message(message='Список авторов:\n' + str(db.sql_read_query('SELECT * FROM author')),
                               user_id=event.obj["message"]["from_id"])
@@ -132,7 +131,6 @@ class VkBot:
         _, author_personal_id = (event.obj["message"]["text"].split())
         author_id = self.vk_session.method(method="users.get", values={"user_ids": author_personal_id})[0]["id"]
 
-        # check if user is admin
         if db.user_role_check(event.obj["message"]["from_id"], "admin") and db.user_role_check(author_id, "author"):
             db.sql_execute_query(f'DELETE FROM author WHERE author_id = {author_id}')
             self.send_message(message=f'Автор {author_personal_id} был удален!',
@@ -153,7 +151,6 @@ class VkBot:
         _, admin_personal_id = (event.obj["message"]["text"].split())
         admin_id = self.vk_session.method(method="users.get", values={"user_ids": admin_personal_id})[0]["id"]
 
-        # check if user is admin
         if db.user_role_check(event.obj["message"]["from_id"], "admin") and \
                 not db.user_role_check(admin_id, "admin"):
             db.sql_execute_query(f'INSERT INTO admin VALUES ({admin_id})')
@@ -168,7 +165,6 @@ class VkBot:
                                   user_id=event.obj["message"]["from_id"])
 
     def get_admins(self, event: vk_api.bot_longpoll.VkBotMessageEvent, db: DataBase) -> None:
-        # check if user is admin
         if db.user_role_check(event.obj["message"]["from_id"], "admin"):
             self.send_message(message='Список администраторов:\n' + str(db.sql_read_query('SELECT * FROM admin')),
                               user_id=event.obj["message"]["from_id"])
@@ -183,7 +179,6 @@ class VkBot:
         _, admin_personal_id = (event.obj["message"]["text"].split())
         admin_id = self.vk_session.method(method="users.get", values={"user_ids": admin_personal_id})[0]["id"]
 
-        # check if user is admin
         if db.user_role_check(event.obj["message"]["from_id"], "admin") and db.user_role_check(admin_id, "admin"):
             db.sql_execute_query(f'DELETE FROM admin WHERE admin_id = {admin_id}')
             self.send_message(message=f'Администратор {admin_personal_id} был удален!',
@@ -196,67 +191,80 @@ class VkBot:
                 self.send_message(message=f'У вас недостаточно прав!',
                                   user_id=event.obj["message"]["from_id"])
 
-    # TODO: add handling of multiply connection. Validate that only client can request connection
     def request_connection(self, event: vk_api.bot_longpoll.VkBotMessageEvent, db: DataBase) -> None:
         if len(event.obj["message"]["text"].split()) != 2:
             self.invalid_command(event, '/req_connection id_клиента')
             return
 
-        _, client_id = event.obj["message"]["text"].split()
-        client_first_name, client_last_name = self.get_name_from_id(client_id)
-        author_id = event.obj["message"]["from_id"]
-        author_first_name, author_last_name = self.get_name_from_id(author_id)
+        if db.user_role_check(event.obj["message"]["from_id"], 'author'):
+            _, client_id = event.obj["message"]["text"].split()
+            client_first_name, client_last_name = self.get_name_from_id(client_id)
+            author_id = event.obj["message"]["from_id"]
+            author_first_name, author_last_name = self.get_name_from_id(author_id)
 
-        keyboard = VkKeyboard(one_time=True)
-        keyboard.add_button(
-            label='/disconnect',
-            color=VkKeyboardColor.NEGATIVE
-        )
-        if db.is_connect_requested(client_id):
+            keyboard = VkKeyboard(one_time=True)
+            keyboard.add_button(
+                label='/disconnect',
+                color=VkKeyboardColor.NEGATIVE
+            )
+            if db.is_connection_exist(client_id):
+                self.forward_message(
+                    message='У клиента уже есть активное/запрошенное соединение!',
+                    user_id=author_id,
+                    keyboard=keyboard.get_empty_keyboard(),
+                    attachments=[]
+                )
+                return
+
+            if db.is_connection_exist(author_id):
+                self.forward_message(
+                    message='У вас уже есть активное/запрошенное соединение!',
+                    user_id=author_id,
+                    keyboard=keyboard.get_keyboard(),
+                    attachments=[]
+                )
+                return
+
+            db.sql_execute_query(
+                f'INSERT INTO connection(client_id, author_id, answered) VALUES({client_id}, {author_id}, 0)')
+            connection_id = db.sql_read_query(
+                f'SELECT connection_id FROM connection WHERE client_id={client_id} AND author_id={author_id}')[0][0]
+
             self.forward_message(
-                message='У клиента уже есть активное соединение!',
+                message='Запрос на соединение с клиентом отправлен!',
                 user_id=author_id,
                 keyboard=keyboard.get_keyboard(),
                 attachments=[]
             )
-            return
-        elif db.is_connect_requested(author_id):
-            self.forward_message(
-                message='У вас уже есть активное соединение!',
-                user_id=author_id,
-                keyboard=keyboard.get_keyboard(),
-                attachments=[]
+
+            admins_id = db.sql_read_query('SELECT admin_id FROM admin')
+            inline_keyboard = VkKeyboard(
+                one_time=False,
+                inline=True
             )
-            return
-
-        db.sql_execute_query(
-            f'INSERT INTO connection(client_id, author_id, answered) VALUES({client_id}, {author_id}, 0)')
-        connection_id = db.sql_read_query(
-            f'SELECT connection_id FROM connection WHERE client_id={client_id} AND author_id={author_id}')[0][0]
-
-        admins_id = db.sql_read_query('SELECT admin_id FROM admin')
-        inline_keyboard = VkKeyboard(
-            one_time=False,
-            inline=True
-        )
-        inline_keyboard.add_button(
-            label=f'/accept {connection_id}',
-            color=VkKeyboardColor.POSITIVE,
-        )
-        inline_keyboard.add_button(
-            label=f'/decline {connection_id}',
-            color=VkKeyboardColor.NEGATIVE,
-        )
-        for admin_id in admins_id:
-            self.forward_message(
-                message=f'{author_last_name} {author_first_name} (id{author_id}) запрашивает соединение с \
-                {client_last_name} {client_first_name} (id{client_id}) #{connection_id}',
-                user_id=admin_id,
-                keyboard=inline_keyboard.get_keyboard(),
-                attachments=[]
+            inline_keyboard.add_button(
+                label=f'/accept {connection_id}',
+                color=VkKeyboardColor.POSITIVE,
+            )
+            inline_keyboard.add_button(
+                label=f'/decline {connection_id}',
+                color=VkKeyboardColor.NEGATIVE,
+            )
+            for admin_id in admins_id:
+                self.forward_message(
+                    message=f'{author_last_name} {author_first_name} (id{author_id}) запрашивает соединение с \
+                    {client_last_name} {client_first_name} (id{client_id}) #{connection_id}',
+                    user_id=admin_id,
+                    keyboard=inline_keyboard.get_keyboard(),
+                    attachments=[]
+                )
+        else:
+            self.send_message(
+                message='Только автор может запросить соединение!',
+                user_id=event.obj["message"]["from_id"],
             )
 
-    def accept_connection(self, event: vk_api.bot_longpoll.VkBotMessageEvent, db: DataBase):
+    def accept_connection(self, event: vk_api.bot_longpoll.VkBotMessageEvent, db: DataBase) -> None:
         if len(event.obj["message"]["text"].split()) != 2:
             self.invalid_command(event, '/accept id_соединения')
             return
@@ -272,7 +280,6 @@ class VkBot:
                     message=f"Соединение #{connection_id} уже было одобрено!",
                     user_id=admin_id
                 )
-
             else:
                 db.sql_execute_query(f"UPDATE connection SET answered=1 WHERE connection_id = {connection_id}")
                 keyboard = VkKeyboard(one_time=True)
@@ -298,7 +305,7 @@ class VkBot:
                 user_id=admin_id
             )
 
-    def decline_connection(self, event: vk_api.bot_longpoll.VkBotMessageEvent, db: DataBase):
+    def decline_connection(self, event: vk_api.bot_longpoll.VkBotMessageEvent, db: DataBase) -> None:
         if len(event.obj["message"]["text"].split()) != 2:
             self.invalid_command(event, '/decline id_соединения')
             return
@@ -335,34 +342,36 @@ class VkBot:
         client_id = self.vk_session.method(method="users.get", values={"user_ids": client_personal_id})[0]["id"]
         author_id = self.vk_session.method(method="users.get", values={"user_ids": author_personal_id})[0]["id"]
 
-        # TODO: Add handler of unique connection
-        # check if user is admin
         if db.user_role_check(event.obj["message"]["from_id"], "admin"):
-            keyboard = VkKeyboard(one_time=True)
-            keyboard.add_button('/disconnect', VkKeyboardColor.NEGATIVE)
+            if not db.is_result_exists(f'SELECT client_id, author_id FROM connection WHERE client_id = {client_id} AND author_id = {author_id}'):
+                keyboard = VkKeyboard(one_time=True)
+                keyboard.add_button('/disconnect', VkKeyboardColor.NEGATIVE)
 
-            db.sql_execute_query(
-                f'INSERT INTO connection(client_id, author_id, answered) VALUES({client_id}, {author_id}, 1)')
+                db.sql_execute_query(
+                    f'INSERT INTO connection(client_id, author_id, answered) VALUES({client_id}, {author_id}, 1)')
 
-            self.send_message(message=f'Связь между клиентом: {client_personal_id} и '
-                                      f'автором: {author_personal_id} установлена!',
-                              user_id=event.obj["message"]["from_id"])
+                self.send_message(message=f'Связь между клиентом: {client_personal_id} и '
+                                          f'автором: {author_personal_id} установлена!',
+                                  user_id=event.obj["message"]["from_id"])
 
-            self.forward_message(message='Связь с автором установлена!',
-                                 user_id=client_id,
-                                 keyboard=keyboard.get_keyboard(),
-                                 attachments=event.obj["message"]["attachments"])
+                self.forward_message(message='Связь с автором установлена!',
+                                     user_id=client_id,
+                                     keyboard=keyboard.get_keyboard(),
+                                     attachments=event.obj["message"]["attachments"])
 
-            self.forward_message(message='Связь с клиентом установлена!',
-                                 user_id=author_id,
-                                 keyboard=keyboard.get_keyboard(),
-                                 attachments=event.obj["message"]["attachments"])
+                self.forward_message(message='Связь с клиентом установлена!',
+                                     user_id=author_id,
+                                     keyboard=keyboard.get_keyboard(),
+                                     attachments=event.obj["message"]["attachments"])
+            else:
+                self.send_message(message=f'Связь между клиентом: {client_personal_id} и '
+                                          f'автором: {author_personal_id} уже существует!',
+                                  user_id=event.obj["message"]["from_id"])
         else:
             self.send_message(message=f'У вас недостаточно прав!',
                               user_id=event.obj["message"]["from_id"])
 
     def get_connections(self, event: vk_api.bot_longpoll.VkBotMessageEvent, db: DataBase) -> None:
-        # check if user is admin
         if db.user_role_check(event.obj["message"]["from_id"], "admin"):
             self.send_message(message='Список установленных соединений:\n' +
                                       str(db.sql_read_query('SELECT * FROM connection')),
@@ -380,23 +389,29 @@ class VkBot:
         client_id = self.vk_session.method(method="users.get", values={"user_ids": client_personal_id})[0]["id"]
         author_id = self.vk_session.method(method="users.get", values={"user_ids": author_personal_id})[0]["id"]
 
-        # TODO: Check if connection exists before deletion
-        # check if user is admin
         if db.user_role_check(event.obj["message"]["from_id"], "admin"):
-            db.sql_execute_query(f'DELETE FROM connection WHERE client_id = {client_id} AND author_id = {author_id}')
-            self.send_message(message=f'Связь между клиентом: {client_personal_id} и '
-                                      f'автором: {author_personal_id} прервана!',
-                              user_id=event.obj["message"]["from_id"])
+            if db.is_result_exists(f'SELECT client_id, author_id FROM connection WHERE client_id = {client_id} AND author_id = {author_id}'):
+                db.sql_execute_query(f'DELETE FROM connection WHERE client_id = {client_id} AND author_id = {author_id}')
+                self.send_message(message=f'Связь между клиентом: {client_personal_id} и '
+                                          f'автором: {author_personal_id} прервана!',
+                                  user_id=event.obj["message"]["from_id"])
+            else:
+                self.send_message(message=f'Связи между клиентом: {client_personal_id} и '
+                                          f'автором: {author_personal_id} не существует!',
+                                  user_id=event.obj["message"]["from_id"])
         else:
             self.send_message(message=f'У вас недостаточно прав!',
                               user_id=event.obj["message"]["from_id"])
 
-    # TODO: error handle
     def disconnect(self, event: vk_api.bot_longpoll.VkBotMessageEvent, db: DataBase) -> None:
-        db.sql_execute_query(f'DELETE FROM connection WHERE client_id = {event.obj["message"]["from_id"]} OR '
-                             f'author_id = {event.obj["message"]["from_id"]}')
-        self.send_message('Вы отключились от чата!',
-                          user_id=event.obj["message"]["from_id"])
+        if db.is_connection_exist(event.obj["message"]["from_id"]):
+            db.sql_execute_query(f'DELETE FROM connection WHERE client_id = {event.obj["message"]["from_id"]} OR '
+                                 f'author_id = {event.obj["message"]["from_id"]}')
+            self.send_message('Вы отключились от чата!',
+                              user_id=event.obj["message"]["from_id"])
+        else:
+            self.send_message('Вы не подключены к чату!',
+                              user_id=event.obj["message"]["from_id"])
 
     def send_message(self, message: str, user_id: int) -> None:
         self.vk_session.get_api().messages.send(
